+++ src/client/components/router/Router.jsx
@@ -1,7 +1,8 @@
 import React from "react";
 import { BrowserRouter, Switch, Route } from "react-router-dom";
 import Home from "pages/home/Home";
-import Rooms from "pages/rooms/Rooms";
+import Lobbies from "pages/lobbies/Lobbies";
+import Lobby from "pages/lobby/LobbyContainer";
 import Game from "pages/game/Game";
 import { GameContextProvider } from "store";
 
@@ -16,12 +17,18 @@ export default function Router() {
     <BrowserRouter>
       <Switch>
         <Route exact path="/" component={Home} />
-        <Route exact path="/rooms" component={Rooms} />
+
+        <Route path="/rooms">
+          <Lobbies />
+          <Route exact path="/rooms/id" component={Lobby} />
+        </Route>
+
         <Route path="/game">
           <GameContextProvider>
             <Game />
           </GameContextProvider>
         </Route>
+
         <Route render={() => <h1>Not found!</h1>} />
       </Switch>
     </BrowserRouter>
