+++ src/client/pages/lobbies/Lobbies.scss
@@ -9,10 +9,6 @@
   scrollbar-width: none; /* Firefox */
 }
 
-.word-breaker {
-  word-break: break-word;
-}
-
 .create-modal {
   @apply flex flex-col justify-center items-center bg-white rounded py-10 px-20;
 
@@ -22,5 +18,13 @@
 
   h2 {
     @apply text-2xl my-2 font-bold;
+  }
+}
+
+.lobby-item {
+  @apply border border-grey-200 justify-between items-center rounded-lg shadow-md p-3 mb-3;
+
+  .status-dot {
+    @apply h-5 w-5 rounded-md mr-2;
   }
 }
